pyproject.toml
[project]
name = "mtblsws-py"
version = "2.2.0"
description = ""
authors = [{ name = "MetaboLights Team", email = "metabolights-dev@ebi.ac.uk" }]
requires-python = ">=3.12,<4"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.6,<3",
    "pydantic-settings>=2.8.1,<3",
    "flask-restful>=0.3.10,<0.4",
    "toml>=0.10.2,<0.11",
    "metaspace2020>=2.0.9,<3",
    "flask-cors>=5.0.1",
    "flask-mail>=0.10.0,<0.11",
    "celery>=5.4.0",
    "redis>=6.4.0,<7",
    "psycopg2>=2.9.10",
    "zeep>=4.3.1,<5",
    "elasticsearch>=1.4.0,<2",
    "passlib>=1.7.4,<2",
    "cirpy>=1.0.2,<2",
    "ctfile>=0.1.12,<0.2",
    "pubchempy>=1.0.4,<2",
    "unidecode>=1.3.8,<2",
    "chembl-structure-pipeline>=1.2.2,<2",
    "rdkit>=2024.9.5",
    "jsonpickle>=4.0.2,<5",
    "gspread>=6.2.0,<7",
    "owlready2>=0.47,<0.48",
    "gspread-dataframe>=4.0.0,<5",
    "oauth2client>=4.1.3,<5",
    "google-api-python-client>=2.163.0,<3",
    "jira>=3.8.0,<4",
    "boto3>=1.37.8,<2",
    "xmltodict>=0.14.2",
    "cascadict>=0.8.5,<0.9",
    "fuzzywuzzy>=0.18.0,<0.19",
    "marshmallow==2.21.0",
    "gunicorn>=23.0.0,<24",
    "charset-normalizer==3.4.1",
    "email-validator>=2.2.0,<3",
    "metabolights-utils>=1.4.8,<2",
    "python-levenshtein>=0.27.1,<0.28",
    "pyyaml>=6.0.3,<7",
    "pyjwt>=2.10.1,<3",
    "flask-restful-swagger",
    "isatools",
    "sqlalchemy==1.4.54",
    "chardet>=5.2.0",
    "pronto>=2.7.2",
    "mzml2isa>=1.1.1",
    "python-keycloak>=5.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "vulture>=2.14,<3",
    "ruff>=0.14.7,<1",
    "import-linter>=2.5,<3",
    "pre-commit>=4.5.0"
]

[tool.uv.sources]
flask-restful-swagger = { git = "https://github.com/EBI-Metabolights/flask-restful-swagger.git", rev = "master" }
isatools = { git = "https://github.com/EBI-Metabolights/isa-api.git", rev = "master" }

[tool.hatch.build.targets.sdist]
include = [
    "app",
    "scripts",
    "tests",
    "resources",
]

[tool.hatch.build.targets.wheel]
include = [
    "app",
    "scripts",
    "tests",
    "resources",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
extend-select = ["E4", "E7", "E9", "F", "I", "G", "SLF",  "Q"]
ignore = [ "ANN401", "F841", "E722", "E731", "SLF001", "G001", "G002", "G003", "G004"]

fixable = ["ALL"]

[tool.importlinter]
include_external_packages = true
root_packages = ["app"]

[[tool.importlinter.contracts]]
name = "Unexpected dependencies"
type = "forbidden"
source_modules = [
   "app",
]
forbidden_modules = [
    "tweepy",
    "jose",
    "pillow"
]
ignore_imports = [
]
